+++ nucleoid/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.2
+- Add CSS property: flex, flexBasis, flexShrink
+- Add HTML tag: hr
+- Add HTML attribute: role
+- Refactoring CSS: BorderCSS.none, BorderCSS.inherit
+
 ## 0.1.1
 - Add class `Route`
 ```